phase_4_scheduler_scraping/requirements.txt
playwright>=1.44.0
beautifulsoup4>=4.12.3
httpx>=0.27.0
pyyaml>=6.0.1
pytest>=8.0.0
